requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
enum-compat==0.0.3
grpcio==1.50.0
idna==3.4
protobuf==3.20.3
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
ydb==2.10.0
